requirements.txt
openai>=1.40.0
PyYAML>=6.0
